refactor: separate clickOnKbd function fron main.js
